Move contents in search_context.py to __init__.py
