+++ appium/webdriver/webelement.py
@@ -15,9 +15,7 @@
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.command import Command as RemoteCommand
 
-from .extensions.search_context.search_context import (
-    AppiumWebElementSearchContext
-)
+from .extensions.search_context import AppiumWebElementSearchContext
 from .mobilecommand import MobileCommand as Command
 
 # Python 3 imports
